refactor: move CI detection to separate package to be reused.
