+++ ci/doc.go
@@ -1,0 +1,5 @@
+// Copyright 2024 Terramate GmbH
+// SPDX-License-Identifier: MPL-2.0
+
+// Package ci provides helper functions to detect the CI/CD platform.
+package ci
